refactor(dashboard): simplify DashboardPage layout and integrate ProtectedLayout; remove NotFound handling components
