+++ frontend/src/features/products/routes/products.routes.jsx
@@ -1,9 +1,21 @@
 import ProductsListPage from '../pages/ProductsListPage.jsx';
+import { ProtectedRoute } from '@/infrastructure/router/ProtectedRoute.jsx';
+import { ProtectedLayout } from '@/shared/components/layout/ProtectedLayout.jsx';
 
 export const productsRoutes = [
   {
     path: '/products',
-    element: <ProductsListPage />,
+    element: (
+      <ProtectedRoute>
+        <ProtectedLayout />
+      </ProtectedRoute>
+    ),
+    children: [
+      {
+        index: true,
+        element: <ProductsListPage />,
+      },
+    ],
   },
 ];
