+++ frontend/src/infrastructure/router/routes.config.jsx
@@ -1,22 +1,13 @@
 import { authRoutes } from "@/features/auth/routes/auth.routes.jsx";
 import { dashboardRoutes } from "@/features/dashboard/routes/dashboard.routes.jsx";
-import { notfoundRoutes } from "@/pages/routes/notfound.routes";
-import { AuthRedirect } from "./AuthRedirect.jsx";
-import { ProtectedRoute } from "./ProtectedRoute.jsx";
-import { Navigate } from "react-router-dom";
+import { productsRoutes } from "@/features/products/routes/products.routes.jsx";
+import { notfoundRoutes } from "./notfound.routes.jsx";
 import { createBrowserRouter } from "react-router-dom";
 
 const router = [
   ...authRoutes,
   ...dashboardRoutes,
-  {
-    path: "/products",
-    element: (
-      <ProtectedRoute>
-        <Navigate to="/dashboard/products" replace />
-      </ProtectedRoute>
-    ),
-  },
+  ...productsRoutes,
   ...notfoundRoutes
 ];
